t0.0.1.1
remove default SSPWA branding from project code
