+++ README.md
@@ -11,7 +11,7 @@
 
 ## Summary
 
-CryptoCalc is a MIT-licensed open source project built with the [SSPWA Template](https://sspw.app "Interact with the SSPWA Template") using Vue.js, Solidity smart contracts, and Golang microservices. End the frustration over how long it takes to convert one currency to another. The process can be excruciating slow when you are researching digital or crypto currencies. Especially when you must start with one currency and convert it to multiple currencies before getting to the final currency. The CryptoCalc DApp will do that for you on the fly as easily as possible.
+CryptoCalc is a MIT-licensed open source project built with the [SSPWA Template](https://sspw.app "Interact with the SSPWA Template") using Vue.js, Solidity smart contracts, and Golang microservices. End the frustration over how long it takes to convert one currency to another. The process can be excruciating slow when you are researching digital or crypto currencies, especially when you must start with one currency and convert it to multiple currencies before getting to the final currency. The CryptoCalc DApp will do that for you on the fly as easily as possible.
 
 Connect your MetaMask wallet or enter your starting amount and currency to get started. Easily create a chain of currency conversions and calculate the final amount in any currency you choose. You can even add gas fees when converting one crypto or fiat currency to another. Use this app to quickly calculate profit of potential investment opportunities. You will even be able to add stock trades on the NYSE to your chain of conversions.
