+++ cryotocalc/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "SSPWA",
-  "version": "1.1.3",
+  "name": "CryptoCalc",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
