+++ cryotocalc/public/icons/ios/apple-icon-152.png
