+++ cryotocalc/public/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0,maximum-scale=1.0,user-scalable=0">
     <meta name="apple-mobile-web-app-capable" content="yes">
     <link rel="icon" href="<%= BASE_URL %>icons/favicon.ico">
-    <title>Single Screen Progressive Web App</title>
+    <title>CryptoCalc Currency Converter DApp</title>
     <link rel="stylesheet" href="https://use.typekit.net/nfx5sjj.css">
     <link rel="apple-touch-icon" sizes="180x180" href="icons/ios/apple-icon-180.png">
     <link rel="icon" type="image/png" sizes="32x32" href="icons/favicon-32x32.png">
