+++ cryotocalc/src/components/sspwa/appFooter.vue
